+++ src/Blocks/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-duplicates */
 /**
  * The main file for all the gutenberg Blocks.
  *