+++ src/Plugin.php
@@ -56,8 +56,7 @@ final class Plugin {
 	 * @return void
 	 */
 	private function configureSettings(): void {
-		$url = esc_url_raw( wp_unslash( $_SERVER['HTTP_HOST'] ?? '' ) );
-		Settings::setHost( substr( $url, strpos( $url, ':' ) + 3 ) ?: null );
+		Settings::setHost();
 		Settings::setPluginName( plugin_basename( __DIR__ . '/index.php' ) );
 	}